Try fix

+++ backend/src/main/kotlin/ru/nsu/ctf/paketnikback/domain/service/impl/MinioServiceImpl.kt
@@ -258,7 +258,7 @@ class MinioServiceImpl(
                     .builder()
                     .bucket(bucketName)
                     .`object`(fileName)
-                    .stream(inputStream, file.size, -1)
+                    .stream(inputStream, file.size.toLong(), -1)
                     .contentType("Bytes")
                     .build(),
             )
